Make FUniqueNetIdEpic use EOS SDK id types
The EOS SDK has two types of user id, ProductUserId (PUID)
and EpicAccountId (EAID). Depending on the endpoint one of
the two has to be used.
The new implementation uses the PUID, which is always available
as the primary source for the new FUniqueNetIdEpic and stores
the EAID alongside it, to be retrieved as needed.