Fix casting bug with FUniqueNetId
Previously when casting an FUniqueNetId to an FUniqueNetIdEpic
the explicit casting operator would only consider the PUID of
the FUniqueNetId, discarding any EAID.
To fix this, the GetBytes() and GetSize() methods had to be changed.
GetSize() now always returns the maximum size of a PUID and,
if the field is valid, the maximum size of an EAID.
GetBytes() returns a byte representation of the PUID and, if the field
is valid the byte representation of the EAID.
Note that both of them are stored in the same array. The PUID is stored
at index 0, the EAID, stored at EOS_PRODUCTUSERID_MAX_LENGTH.