Fix Session creation (#15)
* Remove duplicate SessionModificationHandle creation

* Removed unnecessary changes

* Add CHANGELOG

* Update CHANGELOG

* Add git changelog generator config

* Add templates for Issues and PullRequests

* Update issue templates

* Make address of DevAuthTool customizable (#24)

* Make address of DevAuthTool customizable
* Update Documentation

* Update README.md

* Update README.md

* Update CHANGELOG

* Fix missing devToolPort initialization

Fixes #22 - The constructor for the OnlineSubsystemEpic class now
properly initializes the DevToolAddress field

* Update CHANGELOG

* Fix casting bug with FUniqueNetId

Previously when casting an FUniqueNetId to an FUniqueNetIdEpic
the explicit casting operator would only consider the PUID of
the FUniqueNetId, discarding any EAID.
To fix this, the GetBytes() and GetSize() methods had to be changed.
GetSize() now always returns the maximum size of a PUID and,
if the field is valid, the maximum size of an EAID.
GetBytes() returns a byte representation of the PUID and, if the field
is valid the byte representation of the EAID.
Note that both of them are stored in the same array. The PUID is stored
at index 0, the EAID, stored at EOS_PRODUCTUSERID_MAX_LENGTH.

* Fix session creation always returning NotFound

Calling UpdateSessionModification was unnecessary to change the session
locally before sending it to the backend. It is only needed if an
already existing session needs to be modified.

* Fix Session not starting

After a session has been created the status wasn't set from "Creating"
to "Pending" resulting in the session never starting.

* Fix session creation

Fix: A created sesion now gets a proper id
Changed: Inlined methods only used once

* Remove duplicate SessionModificationHandle creation

* Removed unnecessary changes

* Fix session creation always returning NotFound

Calling UpdateSessionModification was unnecessary to change the session
locally before sending it to the backend. It is only needed if an
already existing session needs to be modified.

* Fix Session not starting

After a session has been created the status wasn't set from "Creating"
to "Pending" resulting in the session never starting.

* Fix session creation

Fix: A created sesion now gets a proper id
Changed: Inlined methods only used once

* Remove const from UpdateSessionSearchParameters

Fix: UpdateSessionSearchParameters had a wrong const on the
EOS_HSessionSearch parameter

* Remove EOS_HSessionSearch being a smart pointer

Fix: EOS_HSessionSearch being a smart pointer resulted in crashes
when the pointer was later released through SDK methods.

* Add support for connect login and support for external account creation (#29)

* Add EAS: to BP login node

* Add framework for "connect -> create user"
Added: The identity interface now has a framework for creating a user
when the login callback received a continuance token.
This is currently not implemented fully as there's no way to convert
a continuance token from an to strings.

* Redesign Blueprint login node
Fixed/Changed: Using the UEpicOnlineServicesLoginTask now always
logs in with an epic account
Added: UEpicOnlineServicesConnectLoginTask can now be used to log in
with any supported external account provider.

Fixes: #27

* Prepare FUniqueNetIdEpic for friends interface

Fix: Various session related net ids were still using the
FUniqueNetIdEpic type instead of FUniqueNetIdString
Changed: Fixed typo in FUniqueNetIdEpic::ToProductUserId
Removed: Constructing an FUniqueNetIdEpic from a string is not
possible anymore since it's ambiguous which Id (PUID or EAID) is meant.
The caller should first create a PUID or EAID via the static methods in
FUniqueNetIdEpic and then construct the unique net id itself.
Changed: GetBytes now includes a single unsigned integer at index 0
indicating if the byte array contains a PUID, EAID or both as this
can't be determined by size anymore.
Added: Constructing an FUniqueNetIdEpic from an EAID is now possible
Added: IsProductUserId
Changed: IsValid now returns whether either of the internal ids is valid
Changed: PUID and EAID displayed by ToDebugString are now properly
redacted in non debug builds
Changed: FUniqueNetIdEpic::GetHash now calculates an improved hash

* Fix user info query

* Compiled code with fixed identity and user interface calls. Will need to test soon.

* Found a bug where the query was never being triggered, so one could never see the names being printed. Tried to fix it by adding more attributes and updating the tuple, which should work but the actual display name is never being returned. Have to figure that out later.

* Queries now complete properly - first problem fix. Am having trouble receiving the names of queried people now though.

* Fixed the final error in user info queries - which was getting the name displayed of the user. Turns out there was a problem in OnlineSubsystemEpicTypes.h that was using the product user id instead of the account user id.

* Commented out presence as it throws a compile error.

* Did cleanup of code and removed friends interface from here for the sole purpose of demonstrating the bug fix for this issue.

* Addressed all commments for PR besides merge conflict.

* Fix file spelling error

* FIx minor style things

Co-authored-by: MikhailSorokin <sorokinmike@yahoo.com>
Co-authored-by: Abrar <afs167@gmail.com>
Co-authored-by: Robert Logiewa <Ruhrpottpatriot@gmx.de>

* Remove duplicate SessionModificationHandle creation

* Removed unnecessary changes

* Fix session creation always returning NotFound

Calling UpdateSessionModification was unnecessary to change the session
locally before sending it to the backend. It is only needed if an
already existing session needs to be modified.

* Fix Session not starting

After a session has been created the status wasn't set from "Creating"
to "Pending" resulting in the session never starting.

* Fix session creation

Fix: A created sesion now gets a proper id
Changed: Inlined methods only used once

* Remove const from UpdateSessionSearchParameters

Fix: UpdateSessionSearchParameters had a wrong const on the
EOS_HSessionSearch parameter

* Remove EOS_HSessionSearch being a smart pointer

Fix: EOS_HSessionSearch being a smart pointer resulted in crashes
when the pointer was later released through SDK methods.

* Remove duplicate SessionModificationHandle creation

* Removed unnecessary changes

* Fix session creation always returning NotFound

Calling UpdateSessionModification was unnecessary to change the session
locally before sending it to the backend. It is only needed if an
already existing session needs to be modified.

* Fix Session not starting

After a session has been created the status wasn't set from "Creating"
to "Pending" resulting in the session never starting.

* Fix session creation

Fix: A created sesion now gets a proper id
Changed: Inlined methods only used once

* Fix minor things

Co-authored-by: Robert Logiewa <Ruhrpottpatriot@gmx.de>
Co-authored-by: MikhailSorokin <sorokinmike96@gmail.com>
Co-authored-by: MikhailSorokin <sorokinmike@yahoo.com>
Co-authored-by: Abrar <afs167@gmail.com>